Additional changes to fix minor service support stuff and increase test coverage. Also making validate token call available using service admin tokens

+++ keystone/backends/sqlalchemy/api/endpoint_template.py
@@ -205,7 +205,8 @@ class EndpointTemplateAPI(BaseEndpointTemplateAPI):
             session = get_session()
         with session.begin():
             endpoints = self.endpoint_get(id, session)
-            session.delete(endpoints)
+            if endpoints:
+                session.delete(endpoints)
 
 
 def get():
